Python: Refactor rescale function
