+++ neural_processes/utils.py
@@ -57,23 +57,6 @@ def fit(niter, save_iter, np, opt, train_set, query_test):
     return mu, sigma
 
 
-def rescale(x, y, dev):
-    """
-    rescale coordinates to range [0, 28*28)
-    rescale y to [0, 1]
-    """
-    scale_x = 28 * 28 - 1
-
-    new_y = y + .5
-    new_y = new_y.to(dev)
-
-    new_x = (x + 1).div(2) * scale_x
-    new_x = new_x.round()
-    new_x = new_x.long().to(dev)
-
-    return new_x, new_y
-
-
 def map_to_img(xc, yc, xt, yt, dev):
     """
     input is of shape as NP input excluding batch_dim, s.t. [num_points, [row_idx, col_idx]]
